Expose SPIRVToOCL12, SPIRVToOCL20 interfaces in SPIRVToOCL.h
