+++ lib/SPIRV/SPIRVToOCL.h
@@ -38,9 +38,13 @@
 ///
 //===----------------------------------------------------------------------===//
 
+#ifndef SPIRVTOOCL_H
+#define SPIRVTOOCL_H
+
 #include "OCLUtil.h"
 #include "SPIRVInternal.h"
 #include "llvm/IR/InstVisitor.h"
+#include "llvm/IR/PassManager.h"
 #include "llvm/Pass.h"
 
 #include <string>
@@ -185,4 +189,142 @@ public:
   bool runOnModule(Module &M) override = 0;
 };
 
+class SPIRVToOCL12Base : public SPIRVToOCLBase {
+public:
+  bool runSPIRVToOCL(Module &M) override;
+
+  /// Transform __spirv_MemoryBarrier to atomic_work_item_fence.
+  ///   __spirv_MemoryBarrier(scope, sema) =>
+  ///       atomic_work_item_fence(flag(sema), order(sema), map(scope))
+  void visitCallSPIRVMemoryBarrier(CallInst *CI) override;
+
+  /// Transform __spirv_ControlBarrier to barrier.
+  ///   __spirv_ControlBarrier(execScope, memScope, sema) =>
+  ///       barrier(flag(sema))
+  void visitCallSPIRVControlBarrier(CallInst *CI) override;
+
+  /// Transform __spirv_OpAtomic functions. It firstly conduct generic
+  /// mutations for all builtins and then mutate some of them seperately
+  Instruction *visitCallSPIRVAtomicBuiltin(CallInst *CI, Op OC) override;
+
+  /// Transform __spirv_OpAtomicIIncrement / OpAtomicIDecrement to
+  /// atomic_inc / atomic_dec
+  Instruction *visitCallSPIRVAtomicIncDec(CallInst *CI, Op OC) override;
+
+  /// Transform __spirv_OpAtomicUMin/SMin/UMax/SMax into
+  /// atomic_min/atomic_max, as there is no distinction in OpenCL 1.2
+  /// between signed and unsigned version of those functions
+  Instruction *visitCallSPIRVAtomicUMinUMax(CallInst *CI, Op OC);
+
+  /// Transform __spirv_OpAtomicLoad to atomic_add(*ptr, 0)
+  Instruction *visitCallSPIRVAtomicLoad(CallInst *CI);
+
+  /// Transform __spirv_OpAtomicStore to atomic_xchg(*ptr, value)
+  Instruction *visitCallSPIRVAtomicStore(CallInst *CI);
+
+  /// Transform __spirv_OpAtomicFlagClear to atomic_xchg(*ptr, 0)
+  /// with ignoring the result
+  Instruction *visitCallSPIRVAtomicFlagClear(CallInst *CI);
+
+  /// Transform __spirv_OpAtomicFlagTestAndTest to
+  /// (bool)atomic_xchg(*ptr, 1)
+  Instruction *visitCallSPIRVAtomicFlagTestAndSet(CallInst *CI);
+
+  /// Transform __spirv_OpAtomicCompareExchange and
+  /// __spirv_OpAtomicCompareExchangeWeak into atomic_cmpxchg. There is no
+  /// weak version of function in OpenCL 1.2
+  Instruction *visitCallSPIRVAtomicCmpExchg(CallInst *CI, Op OC) override;
+
+  /// Conduct generic mutations for all atomic builtins
+  CallInst *mutateCommonAtomicArguments(CallInst *CI, Op OC) override;
+
+  /// Transform atomic builtin name into correct ocl-dependent name
+  Instruction *mutateAtomicName(CallInst *CI, Op OC) override;
+
+  /// Transform SPIR-V atomic instruction opcode into OpenCL 1.2 builtin name.
+  /// Depending on the type, the return name starts with "atomic_" for 32-bit
+  /// types or with "atom_" for 64-bit types, as specified by
+  /// cl_khr_int64_base_atomics and cl_khr_int64_extended_atomics extensions.
+  std::string mapAtomicName(Op OC, Type *Ty);
+};
+
+class SPIRVToOCL12Pass : public llvm::PassInfoMixin<SPIRVToOCL12Pass>,
+                         public SPIRVToOCL12Base {
+public:
+  llvm::PreservedAnalyses run(llvm::Module &M,
+                              llvm::ModuleAnalysisManager &MAM) {
+    return runSPIRVToOCL(M) ? llvm::PreservedAnalyses::none()
+                            : llvm::PreservedAnalyses::all();
+  }
+};
+
+class SPIRVToOCL12Legacy : public SPIRVToOCL12Base, public SPIRVToOCLLegacy {
+public:
+  SPIRVToOCL12Legacy() : SPIRVToOCLLegacy(ID) {
+    initializeSPIRVToOCL12LegacyPass(*PassRegistry::getPassRegistry());
+  }
+  bool runOnModule(Module &M) override;
+
+  static char ID;
+};
+
+class SPIRVToOCL20Base : public SPIRVToOCLBase {
+public:
+  bool runSPIRVToOCL(Module &M) override;
+
+  /// Transform __spirv_MemoryBarrier to atomic_work_item_fence.
+  ///   __spirv_MemoryBarrier(scope, sema) =>
+  ///       atomic_work_item_fence(flag(sema), order(sema), map(scope))
+  void visitCallSPIRVMemoryBarrier(CallInst *CI) override;
+
+  /// Transform __spirv_ControlBarrier to work_group_barrier/sub_group_barrier.
+  /// If execution scope is ScopeWorkgroup:
+  ///    __spirv_ControlBarrier(execScope, memScope, sema) =>
+  ///         work_group_barrier(flag(sema), map(memScope))
+  /// Otherwise:
+  ///    __spirv_ControlBarrier(execScope, memScope, sema) =>
+  ///         sub_group_barrier(flag(sema), map(memScope))
+  void visitCallSPIRVControlBarrier(CallInst *CI) override;
+
+  /// Transform __spirv_Atomic* to atomic_*.
+  ///   __spirv_Atomic*(atomic_op, scope, sema, ops, ...) =>
+  ///      atomic_*(generic atomic_op, ops, ..., order(sema), map(scope))
+  Instruction *visitCallSPIRVAtomicBuiltin(CallInst *CI, Op OC) override;
+
+  /// Transform __spirv_OpAtomicIIncrement / OpAtomicIDecrement to
+  /// atomic_fetch_add_explicit / atomic_fetch_sub_explicit
+  Instruction *visitCallSPIRVAtomicIncDec(CallInst *CI, Op OC) override;
+
+  /// Conduct generic mutations for all atomic builtins
+  CallInst *mutateCommonAtomicArguments(CallInst *CI, Op OC) override;
+
+  /// Transform atomic builtin name into correct ocl-dependent name
+  Instruction *mutateAtomicName(CallInst *CI, Op OC) override;
+
+  /// Transform __spirv_OpAtomicCompareExchange/Weak into
+  /// compare_exchange_strong/weak_explicit
+  Instruction *visitCallSPIRVAtomicCmpExchg(CallInst *CI, Op OC) override;
+};
+
+class SPIRVToOCL20Pass : public llvm::PassInfoMixin<SPIRVToOCL20Pass>,
+                         public SPIRVToOCL20Base {
+public:
+  llvm::PreservedAnalyses run(llvm::Module &M,
+                              llvm::ModuleAnalysisManager &MAM) {
+    return runSPIRVToOCL(M) ? llvm::PreservedAnalyses::none()
+                            : llvm::PreservedAnalyses::all();
+  }
+};
+
+class SPIRVToOCL20Legacy : public SPIRVToOCLLegacy, public SPIRVToOCL20Base {
+public:
+  SPIRVToOCL20Legacy() : SPIRVToOCLLegacy(ID) {
+    initializeSPIRVToOCL20LegacyPass(*PassRegistry::getPassRegistry());
+  }
+  bool runOnModule(Module &M) override;
+  static char ID;
+};
+
 } // namespace SPIRV
+
+#endif // SPIRVTOOCL_H
